education_web/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
diff-match-patch==20200713
Django==2.2
django-crispy-forms==1.8.0
django-formtools==2.3
django-import-export==0.5.1
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==2.0.0
django-simple-captcha==0.5.17
future==0.15.2
httplib2==0.9.2
idna==3.3
mysqlclient==2.1.0
Pillow==9.1.0
pytz==2022.1
requests==2.27.1
six==1.10.0
sqlparse==0.4.2
tablib==3.2.1
urllib3==1.26.9
XlsxWriter==3.0.3
xlwt==1.3.0
